Add exceptin handling for ProductsController

+++ Controllers/ProductsController.cs
@@ -19,41 +19,117 @@ namespace HPlusSportsAPI.Controllers
             _context = context;
         }
 
+        private bool ProductExists(string id)
+        {
+            return _context.Product.Any(e => e.ProductId == id);
+        }
+
         [HttpGet]
+        [Produces(typeof(DbSet<Product>))]
         public IActionResult GetProduct()
         {
             return new ObjectResult(_context.Product);
         }
 
-        [HttpGet("{id}", Name = "GetProduct")]
+        [HttpGet("{id}")]
         public async Task<IActionResult> GetProduct([FromRoute] string id)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
             var product = await _context.Product.SingleOrDefaultAsync(m => m.ProductId == id);
+
+            if (product == null)
+            {
+                return NotFound();
+            }
+
             return Ok(product);
         }
 
         [HttpPost]
+        [Produces(typeof(Product))]
         public async Task<IActionResult> PostProduct([FromBody] Product product)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
             _context.Product.Add(product);
-            await _context.SaveChangesAsync();
-            return CreatedAtAction("GetProduct", new { id = product.ProductId }, product);
+
+            try
+            {
+                await _context.SaveChangesAsync();
+            }
+            catch (DbUpdateException)
+            {
+                if (ProductExists(product.ProductId))
+                {
+                    return new StatusCodeResult(StatusCodes.Status409Conflict);
+                }
+                else
+                {
+                    throw;
+                }
+            }
+
+                return CreatedAtAction("GetProduct", new { id = product.ProductId }, product);
         }
 
         [HttpPut("{id}")]
-        public async Task <IActionResult> PutProduct([FromRoute] int id, [FromBody] Product product)
+        [Produces(typeof(Product))]
+        public async Task <IActionResult> PutProduct([FromRoute] string id, [FromBody] Product product)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+
+            if (id != product.ProductId)
+            {
+                return BadRequest();
+            }
             _context.Entry(product).State = EntityState.Modified;
-            await _context.SaveChangesAsync();
-            return Ok(product);
+
+            try
+            {
+                await _context.SaveChangesAsync();
+                return Ok(product);
+            }
+            catch(DbUpdateException)
+            {
+                if (!ProductExists(id))
+                {
+                    return NotFound();
+                }
+                else
+                {
+                    throw;
+                }
+
+            }
         }
 
         [HttpDelete("{id}")]
+        [Produces(typeof(Product))]
         public async Task<IActionResult> DeleteProduct([FromRoute] string id)
         {
+            if(!ModelState.IsValid)
+            {
+                return BadRequest(ModelState);
+            }
+       
             var product = await _context.Product.SingleOrDefaultAsync(m => m.ProductId == id);
+            if(product == null)
+            {
+                return NotFound();
+            }
+
             _context.Product.Remove(product);
             await _context.SaveChangesAsync();
+
             return Ok(product);
         }
     }
